requirements.txt
yfinance
requests
google-generativeai
beautifulsoup4
python-dotenv
